Fix num_frames output type

+++ nemo/collections/tts/modules/fastpitch.py
@@ -284,7 +284,7 @@ class FastPitchModule(NeuralModule):
         # Output FFT
         dec_out, _ = self.decoder(input=len_regulated, seq_lens=dec_lens)
         spect = self.proj(dec_out).transpose(1, 2)
-        num_frames = durs_predicted.sum(axis=-1)
+        num_frames = durs_predicted.sum(axis=-1).to(dtype=torch.int64)
         return (
             spect,
             num_frames,
